# avoid temporary vectors