--- a/REQUIRE
+++ b/REQUIRE
@@ -3,3 +3,4 @@ Optim
 Calculus
 DualNumbers
 Compat
+Distances