--- a/src/NLsolve.jl
+++ b/src/NLsolve.jl
@@ -1,5 +1,6 @@
 module NLsolve
 
+using Distances
 using Optim
 
 import Base.show,
@@ -244,7 +245,7 @@ function assess_convergence(x::Vector,
                             ftol::Real)
     x_converged, f_converged = false, false
 
-    if norm(x - x_previous, Inf) < xtol
+    if !any(isnan, x_previous) && chebyshev(x, x_previous) < xtol
         x_converged = true
     end
 
@@ -257,6 +258,7 @@ function assess_convergence(x::Vector,
     return x_converged, f_converged, converged
 end
 
+include("utils.jl")
 include("newton.jl")
 include("trust_region.jl")
 include("autodiff.jl")