--- a/test/runtests.jl
+++ b/test/runtests.jl
@@ -13,7 +13,13 @@ tests = ["2by2.jl",
 
 println("Running tests:")
 
-for t in tests
-    println(" * $(t)")
-    include(t)
+for test in tests
+    try
+        include(test)
+        println("\t\033[1m\033[32mPASSED\033[0m: $(test)")
+     catch e
+        println("\t\033[1m\033[31mFAILED\033[0m: $(test)")
+        showerror(STDOUT, e, backtrace())
+        rethrow(e)
+     end
 end