… test/2by2.jl