Simple fixes to avoid temporary vectors + adapt the norm computation to sparse matrices. This makes the algorithm an order of magnitude faster for sparse matrices.