@matthieugomez 2015-09-15
I have just updated the pull request — code is now simpler.